Add missing `dir` to the `no-console`
We noticed this was missing, allowing workarounds. I know this repo is in hiatus now, but even if you do not merge, it could be useful for people stumbling over this wondering why this does not trigger. We are actually going to override the rule by removing the list of methods, disallowing all `console` methods (and that would be my preferred option for the recommended ruleset, even).